moved component evaluation methods to estimates class
